+++ src/main/java/com/github/tommyettinger/anim8/PaletteReducer.java
@@ -3285,9 +3285,9 @@ public class PaletteReducer {
         float rdiff, gdiff, bdiff;
         float er, eg, eb;
         byte paletteIndex;
-        float w1 = (float) (24.0 * Math.sqrt(ditherStrength) * populationBias * populationBias * populationBias * populationBias), w3 = w1 * 3f, w5 = w1 * 5f, w7 = w1 * 7f,
-                strength = (float) (0.35 * ditherStrength / (populationBias * populationBias * populationBias * populationBias)),
-                limit = 5f + 90f / (float)Math.sqrt(colorCount+1.5), dmul = (float) (0x1p-8 / populationBias);
+        float w1 = 24f * (float) Math.sqrt(ditherStrength) * populationBias * populationBias * populationBias * populationBias, w3 = w1 * 3f, w5 = w1 * 5f, w7 = w1 * 7f,
+                strength = 0.35f * ditherStrength / (populationBias * populationBias * populationBias * populationBias),
+                limit = 5f + 90f / (float)Math.sqrt(colorCount+1.5f), dmul = 0x1p-8f / populationBias;
 
         for (int py = 0; py < h; py++) {
             int ny = py + 1;
@@ -3316,6 +3316,7 @@ public class PaletteReducer {
                                     | ((bb >>> 3))];
                     used = paletteArray[paletteIndex & 0xFF];
                     pixmap.drawPixel(px, py, used);
+
                     rdiff = (dmul * ((color>>>24)-    (used>>>24))    );
                     gdiff = (dmul * ((color>>>16&255)-(used>>>16&255)));
                     bdiff = (dmul * ((color>>>8&255)- (used>>>8&255)) );
